+++ src/main/java/ru/yandex/practicum/filmorate/service/UserService.java
@@ -46,6 +46,9 @@ public class UserService {
     }
 
     public void deleteUserById(Long id) {
+        if (!userStorage.existsById(id)){
+            throw new NotFoundException("Пользователь с id = {0} не найден.", id);
+        }
         userStorage.deleteById(id);
     }
